Process creation time for files
Ran into a weird bug where casting a single byte to an int would
left-pad the 32 bits with 1's causing totally incorrect values. This
only occurred for values with leading ones which would be negative.
Changed the buffer the disk image is read into to unsigned char* to
resolve the issue.